refactor: extract PreviewBar management to src/content/previewBarManager.ts
Extract preview bar related functions from content.ts into a dedicated module:
- createPreviewBar()
- updatePreviewBar()
- checkPreviewbarState()
- selectSegment()
- updateActiveSegment()
- showTimeWithoutSkips()
- removeDurationAfterSkip()
- durationID constant

Uses dependency injection via initPreviewBarManager() for voteAsync and
updateVisibilityOfPlayerControlsButton.
